+++ lib/weathervision/cli.rb
@@ -2,7 +2,7 @@ module Weathervision
   class CLI < Thor
 
     desc "parse [options]", "Parse weather data" 
-    method_option :mode, :type => :string, :required => true, :desc => "Mode: use 'vtext', 'htext', 'vimage' or 'himage'", :default => 'himage'
+    method_option :mode, :type => :string, :required => true, :desc => "Mode: use 'vimage' or 'himage for vertical or horizontal layout'", :default => 'vimage'
     method_option :apikey, :type => :string, :required => true, :desc => "wunderground.com json API key"
     method_option :current_location, :type => :string, :required => true, :desc => "wunderground.com current weather location query string", :default => 'vlist'
     method_option :forecast_location, :type => :string, :required => true, :desc => "wunderground.com forecast weather location query string", :default => 'vlist'