+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-16-12-2013
+02-05-2013
 ==========
 
 Version 2.0
 -------------
 - Rewrote the parsing of weather data: now using the wunderground.com JSON free developer API. 
-- Added another smaller conky template for correct displaying when using a horizontal conky layout, instead of the more common vertical layout.
-- Added a language selection option.
+- Added another template for a horizontal conky layout
+- Removed the font-based templates and functionality
 
 25-07-2011
 ==========